reports/requirements.txt
boto3
psycopg2
fpdf
python-dotenv
matplotlib
tempfile
base64